+++ src/payementManager.js
@@ -15,7 +15,7 @@ async function getPayement() {
 async function addPayement(order_id, date_payement, amount, payement_method){
     const connection = await pool.getConnection()
     try {
-        const [result] = await connection.execute('INSERT INTO payements (order_id, date_payement, amount, payement_method) VALUE (?, ?, ?, ?)', [order_id, date_payement, amount, payement_method])
+        const [result] = await connection.execute('INSERT INTO payements (order_id, date_payement, amount, payement_method) VALUES (?, ?, ?, ?)', [order_id, date_payement, amount, payement_method])
         return result.insertId
     } catch (error) {
         throw error
@@ -25,31 +25,43 @@ async function addPayement(order_id, date_payement, amount, payement_method){
 } 
 
 async function updatePayement(id, order_id, date_payement, amount, payement_method) {
-    const connection = await pool.getConnection()
+    const connection = await pool.getConnection();
     try {
-        const [result] = await connection.execute('UPDATE payements SET  order_id = ?, date_payement = ?, amount = ?, payement_method = ?', [ order_id, date_payement, amount, payement_method, id,])
-        return result.affectedRows
+        // Vérifier si l'ID du paiement existe
+        const [payementExists] = await connection.execute('SELECT 1 FROM payements WHERE id = ?', [id]);
+        if (payementExists.length === 0) {
+            throw new Error(`Paiement avec l'ID ${id} introuvable.`);
+        }
+
+        const [result] = await connection.execute(
+            'UPDATE payements SET order_id = ?, date_payement = ?, amount = ?, payement_method = ? WHERE id = ?',
+            [order_id, date_payement, amount, payement_method, id]
+        );
+        return result.affectedRows;
     } catch (error) {
-        throw error
+        throw error;
     } finally {
         connection.release();
     }
-    
 }
 
 async function destroyPayement(id) {
-    const connection = await pool.getConnection()
+    const connection = await pool.getConnection();
     try {
-        const [result] = await connection.execute('DELETE FROM payements WHERE id = ?', [id])
-        return result.affectedRows
+        // Vérifier si l'ID du paiement existe
+        const [payementExists] = await connection.execute('SELECT 1 FROM payements WHERE id = ?', [id]);
+        if (payementExists.length === 0) {
+            throw new Error(`Paiement avec l'ID ${id} introuvable.`);
+        }
+
+        const [result] = await connection.execute('DELETE FROM payements WHERE id = ?', [id]);
+        return result.affectedRows;
     } catch (error) {
-        throw error
+        throw error;
     } finally {
         connection.release();
     }
 }
-
-
 
 module.exports = {
     getPayement,
