+++ src/productManager.js
@@ -36,29 +36,44 @@ async function addProduct(name, description, price, stock, category, barcode, st
 } 
 
 async function updateProduct(id, name, description, price, stock, category, barcode, status) {
-    const connection = await pool.getConnection()
+    const connection = await pool.getConnection();
     try {
-        const [result] = await connection.execute('UPDATE products SET  name = ?, description = ?, price = ?, stock = ?, category = ?, barcode = ?, status = ? WHERE id = ?', [ name, description, price, stock, category, barcode, status, id,])
-        return result.affectedRows
+        // Vérifier si l'ID du produit existe
+        const [productExists] = await connection.execute('SELECT 1 FROM products WHERE id = ?', [id]);
+        if (productExists.length === 0) {
+            throw new Error(`Produit avec l'ID ${id} introuvable.`);
+        }
+
+        const [result] = await connection.execute(
+            'UPDATE products SET name = ?, description = ?, price = ?, stock = ?, category = ?, barcode = ?, status = ? WHERE id = ?',
+            [name, description, price, stock, category, barcode, status, id]
+        );
+        return result.affectedRows;
     } catch (error) {
-        throw error
+        throw error;
     } finally {
         connection.release();
     }
-    
 }
 
 async function destroyProduct(id) {
-    const connection = await pool.getConnection()
+    const connection = await pool.getConnection();
     try {
-        const [result] = await connection.execute('DELETE FROM products WHERE id = ?', [id])
-        return result.affectedRows
+        // Vérifier si l'ID du produit existe
+        const [productExists] = await connection.execute('SELECT 1 FROM products WHERE id = ?', [id]);
+        if (productExists.length === 0) {
+            throw new Error(`Produit avec l'ID ${id} introuvable.`);
+        }
+
+        const [result] = await connection.execute('DELETE FROM products WHERE id = ?', [id]);
+        return result.affectedRows;
     } catch (error) {
-        throw error
+        throw error;
     } finally {
         connection.release();
     }
 }
+
